+++ lib/hubspot_client/model/communication_preference.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module HubspotClient
+  module Model
+    class CommunicationPreference
+      def self.definitions
+        Client::CommunicationPreference.new.definitions
+      end
+    end
+  end
+end