+++ spec/hubspot_client/model/contact_spec.rb
@@ -109,5 +109,31 @@ module HubspotClient
         end
       end
     end
+
+    describe '.create_communication_subscription' do
+      subject(:contact) { described_class.new(email: 'darth.vader@example.com') }
+
+      let(:client_communication_preference) { instance_double(Client::CommunicationPreference) }
+
+      before do
+        allow(Client::CommunicationPreference).to receive(:new).and_return(client_communication_preference)
+        allow(client_communication_preference).to receive(:subscribe)
+      end
+
+      expected_properties = {
+        emailAddress: 'darth.vader@example.com',
+        subscriptionId: 'subscription_id_value',
+        legalBasis: 'legal_basis_value',
+        legalBasisExplanation: 'legal_basis_explanation_value'
+      }
+
+      it 'creates a subscription' do
+        contact.create_communication_subscription(subscription_id: 'subscription_id_value',
+                                                  legal_basis: 'legal_basis_value',
+                                                  legal_basis_explanation: 'legal_basis_explanation_value')
+
+        expect(client_communication_preference).to have_received(:subscribe).with(expected_properties)
+      end
+    end
   end
 end